All notification logic fixed

+++ app/src/main/java/com/example/group12/logic/FacadeFilter.java
@@ -8,9 +8,9 @@ FACADE STRUCTURAL DESIGN PATTERN
 public class FacadeFilter {
 
 
-    public float getDist(float lat1, float lng1, float lat2, float lng2){
+    public double getDist(float lat1, float lng1, float lat2, float lng2){
         FilterJob filter = new FilterJob();
-        float dist = filter.getDistance(lat1,lng1,lat2,lng2);
+        double dist = filter.getDistance(lat1,lng1,lat2,lng2);
 
 
         return dist;
